* improved method calls * renamed `mulle_objc_object_call_variablemethodid_inline` to `mulle_objc_object_call_variable_inline` * renamed `mulle_objc_object_supercall` to `mulle_objc_object_call_super` * `+deinitialize` now also calls protocol classes `+deinitialize` for tidier cleanup * fix that meta rootclasses wrap around to root infraclass again, even if there are protocolclasses * +method calls also wrap arounf to -methods on protocolclasses attached to the root class * NSObject now inherits the most "hardcore" runtime functionality from MulleObjCRuntimeObject * new base class MulleObject enables @dynamic @property storage for categories (and all others) * nicer output of compilation flags in loadinfo * `struct mulle_objc_typeinfo` no longer restricts instance variable arrays to 16 bit * added pre-alpha `_mulle_objc_ivarsignature_is_compatible` for the benefit of MulleObject
fix bug with loadbits not being set properly
fix slow retainCount code
adapt to changes in mulle-thread

+++ test/signature/parse_signature.c
@@ -98,7 +98,7 @@ static int   test_info( char *s)
       printf( "#%d type= %s\n", i, info.type ? info.type : "NULL");
       printf( "#%d pure_type_end= %s\n", i, info.pure_type_end ? info.pure_type_end : "NULL");
       printf( "#%d name= %s\n", i, info.name ? info.name : "NULL");
-      printf( "#%d n_members= %d\n", i, info.n_members);
+      printf( "#%d n_members= %ld\n", i, (long) info.n_members);
       printf( "#%d has_object= %d\n", i, info.has_object);
       printf( "#%d has_object= %d\n", i, info.has_object);
